+++ tqdm/tests/tests_pandas.py
@@ -1,6 +1,6 @@
 from nose.plugins.skip import SkipTest
 
-from tqdm import tqdm
+from tqdm import tqdm, tqdm_pandas
 from tests_tqdm import with_setup, pretest, posttest, StringIO, closing
 
 
@@ -14,7 +14,7 @@ def test_pandas_groupby_apply():
         raise SkipTest
 
     with closing(StringIO()) as our_file:
-        tqdm.pandas(file=our_file, leave=False, ascii=True)
+        tqdm_pandas(tqdm, file=our_file, leave=False, ascii=True)
 
         df = pd.DataFrame(randint(0, 50, (500, 3)))
         df.groupby(0).progress_apply(lambda x: None)
@@ -44,7 +44,7 @@ def test_pandas_apply():
         raise SkipTest
 
     with closing(StringIO()) as our_file:
-        tqdm.pandas(file=our_file, leave=True, ascii=True)
+        tqdm_pandas(tqdm, file=our_file, leave=True, ascii=True)
         df = pd.DataFrame(randint(0, 50, (500, 3)))
         df.progress_apply(lambda x: None)
 
@@ -70,7 +70,7 @@ def test_pandas_map():
         raise SkipTest
 
     with closing(StringIO()) as our_file:
-        tqdm.pandas(file=our_file, leave=True, ascii=True)
+        tqdm_pandas(tqdm, file=our_file, leave=True, ascii=True)
         dfs = pd.DataFrame(randint(0, 50, (500, 3)),
                            columns=list('abc'))
         dfs.a.progress_map(lambda x: None)
@@ -91,7 +91,7 @@ def test_pandas_leave():
 
     with closing(StringIO()) as our_file:
         df = pd.DataFrame(randint(0, 100, (1000, 6)))
-        tqdm.pandas(file=our_file, leave=True, ascii=True)
+        tqdm_pandas(tqdm, file=our_file, leave=True, ascii=True)
         df.groupby(0).progress_apply(lambda x: None)
 
         our_file.seek(0)
@@ -105,7 +105,7 @@ def test_pandas_leave():
 
 @with_setup(pretest, posttest)
 def test_pandas_deprecation():
-    """ Test bar object instance as argument deprecation """
+    """ Test pandas bar object instance as argument deprecation """
     try:
         from numpy.random import randint
         from tqdm import tqdm_pandas
@@ -122,9 +122,9 @@ def test_pandas_deprecation():
         assert "instead of `tqdm_pandas(tqdm(...))`" in our_file.getvalue()
 
     with closing(StringIO()) as our_file:
-        tqdm_pandas(tqdm, file=our_file, leave=False, ascii=True, ncols=20)
+        tqdm.pandas(tqdm, file=our_file, leave=False, ascii=True, ncols=20)
         df = pd.DataFrame(randint(0, 50, (500, 3)))
         df.groupby(0).progress_apply(lambda x: None)
         # Check deprecation message
         assert "TqdmDeprecationWarning" in our_file.getvalue()
-        assert "instead of `tqdm_pandas(tqdm, ...)`" in our_file.getvalue()
+        assert "instead of `tqdm.pandas(...)`" in our_file.getvalue()
